UPD: Refactor panel scoreboard
