+++ src/components/Playground.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import OddCard from './oddx/OddCard.js';
 
-import PlaygroundScore from './PlaygroundScore';
 import PlaygroundWidgets from './PlaygroundWidgets';
 import PlaygroundCollection from './PlaygroundCollection';
 import 'stylesheets/Playground.scss';
@@ -10,7 +9,6 @@ const Playground = () => {
   return (
     <div className="playground">
       <OddCard black />
-      <PlaygroundScore />
       <PlaygroundWidgets />
       <PlaygroundCollection />
     </div>
